tests/wpt/tests/css/css-shapes/shape-functions/polygon-function-round-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing the polygon() function with the round modifier</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-polygon">
<meta name="assert" content="Tests computed value of the round modifier in the polygon() function">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value(
    "shape-outside",
    "polygon(round 1em, 1px 2px, 3em 4em)",
    "polygon(round 16px, 1px 2px, 48px 64px)");
test_computed_value(
    "shape-outside",
    "polygon(evenodd round 6pt, 1px 2px, 3em 4em, 6pt 6%)",
    "polygon(evenodd round 8px, 1px 2px, 48px 64px, 8px 6%)");
test_computed_value(
    "shape-outside",
    "polygon(round 0px, 1px 2px, 3px 4px)",
    "polygon(1px 2px, 3px 4px)");
test_computed_value(
    "shape-outside",
    "polygon(round calc(1em + 2px), 1px 2px, 3px 4px)",
    "polygon(round 18px, 1px 2px, 3px 4px)");
</script>
</body>
</html>
